Add files via upload
There are many stores in which a survey was conducted based on students i.e.
how much they are spending on different kind of purchases like Video games, Indoor games, 
Toys, Books, Gadgets etc. In the data set (Student Survey), Store setting is the column that explains the Type
of location in which the store is present. By using data set (Student Survey), try to extract the
meaningful Insights.